+++ examples/print-sample/components/hello/world/index.tsx
@@ -1,0 +1,6 @@
+import { WorldComponent } from './world.component';
+import { World } from './world.connect';
+import { WorldStyle } from './world.style';
+
+export { WorldComponent, World, WorldStyle }
+