+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # 🖨️ Printer | Changelog
 
+## 1.1.0
+
+- Added `index.tsx` for new components generated with `printer component ${PATH}`.
+
+- Added new command `printer prisma merge`.
+
+- Added husky pre-commit git hooks
+
+- Added `baseUrl` to `tsconfig.json`
+
 ## 1.0.1
 
 - Added support for the Pascal Namespace component generation when using hyphenated file names.